tests/test_mobile.py
"""
Q3 - MOBILE APP TESTING for AI Document Enhancement System.

Tool Used: Appium (Mobile automation framework)

Mobile App Testing Strategy:
  1. Compatibility Testing - across different devices, OS versions, screen sizes
  2. Performance Testing - response times, memory usage, battery consumption
  3. Functionality Testing - core features work correctly on mobile
  4. Network Testing - behavior under different network conditions
  5. Installation Testing - install, update, uninstall workflows
  6. Interrupt Testing - incoming calls, notifications, low battery
"""

import pytest
import time

try:
    from appium import webdriver as appium_webdriver
    from appium.webdriver.common.appiumby import AppiumBy
    from appium.options.android import UiAutomator2Options
    from appium.options.ios import XCUITestOptions
    APPIUM_AVAILABLE = True
except ImportError:
    APPIUM_AVAILABLE = False


# ═══════════════════════════════════════════════════════════════════════
# SECTION A: ANDROID TESTING WITH APPIUM
# ═══════════════════════════════════════════════════════════════════════

@pytest.mark.skipif(not APPIUM_AVAILABLE, reason="Appium not installed")
class TestAndroidApp:
    """Mobile app testing on Android using Appium UiAutomator2."""

    @pytest.fixture(autouse=True)
    def setup_driver(self):
        options = UiAutomator2Options()
        options.platform_name = "Android"
        options.device_name = "Pixel_6_API_33"
        options.app = "/path/to/app-debug.apk"
        options.automation_name = "UiAutomator2"
        options.no_reset = False

        self.driver = appium_webdriver.Remote(
            "http://localhost:4723/wd/hub", options=options
        )
        self.driver.implicitly_wait(15)
        yield
        self.driver.quit()

    # ─── FUNCTIONALITY TESTING ──────────────────────────────────────

    def test_app_launches_successfully(self):
        """TC-MOB-001: App launches and home screen is displayed."""
        home_title = self.driver.find_element(AppiumBy.ID, "com.docenhancer:id/titleText")
        assert home_title.is_displayed()
        assert "Document Enhancer" in home_title.text

    def test_camera_scan_button_exists(self):
        """TC-MOB-002: Camera scan button is accessible on home screen."""
        scan_btn = self.driver.find_element(AppiumBy.ID, "com.docenhancer:id/scanButton")
        assert scan_btn.is_displayed()
        assert scan_btn.is_enabled()

    def test_document_upload_from_gallery(self):
        """TC-MOB-003: User can upload document image from gallery."""
        upload_btn = self.driver.find_element(AppiumBy.ID, "com.docenhancer:id/uploadButton")
        upload_btn.click()

        gallery_option = self.driver.find_element(AppiumBy.XPATH, "//android.widget.TextView[@text='Gallery']")
        gallery_option.click()

        time.sleep(2)
        first_image = self.driver.find_element(AppiumBy.XPATH, "(//android.widget.ImageView)[1]")
        first_image.click()

        preview = self.driver.find_element(AppiumBy.ID, "com.docenhancer:id/previewImage")
        assert preview.is_displayed()

    def test_ocr_processing_displays_result(self):
        """TC-MOB-004: OCR processing completes and shows extracted text."""
        self.driver.find_element(AppiumBy.ID, "com.docenhancer:id/uploadButton").click()
        time.sleep(1)
        self.driver.find_element(AppiumBy.ID, "com.docenhancer:id/processButton").click()

        progress = self.driver.find_element(AppiumBy.ID, "com.docenhancer:id/progressBar")
        assert progress.is_displayed()

        time.sleep(10)  # Wait for OCR
        result_text = self.driver.find_element(AppiumBy.ID, "com.docenhancer:id/resultText")
        assert len(result_text.text) > 0

    def test_navigation_between_screens(self):
        """TC-MOB-005: Navigation between all main screens works."""
        screens = ["Home", "Documents", "History", "Settings"]
        for screen in screens:
            nav_item = self.driver.find_element(AppiumBy.XPATH, f"//android.widget.TextView[@text='{screen}']")
            nav_item.click()
            time.sleep(0.5)
            assert self.driver.find_element(AppiumBy.ID, f"com.docenhancer:id/{screen.lower()}Screen").is_displayed()

    # ─── COMPATIBILITY TESTING ──────────────────────────────────────

    def test_portrait_layout(self):
        """TC-MOB-006: App renders correctly in portrait orientation."""
        self.driver.orientation = "PORTRAIT"
        home = self.driver.find_element(AppiumBy.ID, "com.docenhancer:id/mainLayout")
        size = home.size
        assert size["height"] > size["width"]

    def test_landscape_layout(self):
        """TC-MOB-007: App renders correctly in landscape orientation."""
        self.driver.orientation = "LANDSCAPE"
        home = self.driver.find_element(AppiumBy.ID, "com.docenhancer:id/mainLayout")
        size = home.size
        assert size["width"] > size["height"]

    # ─── PERFORMANCE TESTING ────────────────────────────────────────

    def test_app_launch_time(self):
        """TC-MOB-008: App launches within acceptable time (< 3 seconds)."""
        start_time = time.time()
        self.driver.find_element(AppiumBy.ID, "com.docenhancer:id/titleText")
        launch_time = time.time() - start_time
        assert launch_time < 3.0, f"App took {launch_time:.1f}s to launch (max 3s)"

    def test_document_list_scroll_performance(self):
        """TC-MOB-009: Scrolling through documents list is smooth."""
        self.driver.find_element(AppiumBy.XPATH, "//android.widget.TextView[@text='Documents']").click()
        start = time.time()
        self.driver.swipe(500, 1500, 500, 500, 300)
        scroll_time = time.time() - start
        assert scroll_time < 1.0, "Scrolling was not smooth"

    # ─── NETWORK CONDITION TESTING ──────────────────────────────────

    def test_offline_mode_shows_cached_docs(self):
        """TC-MOB-010: App shows cached documents when offline."""
        self.driver.set_network_connection(0)  # Airplane mode
        time.sleep(1)

        docs_tab = self.driver.find_element(AppiumBy.XPATH, "//android.widget.TextView[@text='Documents']")
        docs_tab.click()

        offline_msg = self.driver.find_element(AppiumBy.ID, "com.docenhancer:id/offlineIndicator")
        assert offline_msg.is_displayed()

        self.driver.set_network_connection(6)  # Restore WiFi + Data


# ═══════════════════════════════════════════════════════════════════════
# SECTION B: iOS TESTING WITH APPIUM
# ═══════════════════════════════════════════════════════════════════════

@pytest.mark.skipif(not APPIUM_AVAILABLE, reason="Appium not installed")
class TestIOSApp:
    """Mobile app testing on iOS using Appium XCUITest."""

    @pytest.fixture(autouse=True)
    def setup_driver(self):
        options = XCUITestOptions()
        options.platform_name = "iOS"
        options.device_name = "iPhone 15 Pro"
        options.platform_version = "17.0"
        options.app = "/path/to/DocEnhancer.app"
        options.automation_name = "XCUITest"

        self.driver = appium_webdriver.Remote(
            "http://localhost:4723/wd/hub", options=options
        )
        self.driver.implicitly_wait(15)
        yield
        self.driver.quit()

    def test_ios_app_launches(self):
        """TC-IOS-001: App launches on iOS device."""
        title = self.driver.find_element(AppiumBy.ACCESSIBILITY_ID, "appTitle")
        assert title.is_displayed()

    def test_ios_camera_permission_prompt(self):
        """TC-IOS-002: Camera permission dialog is shown on first scan."""
        scan_btn = self.driver.find_element(AppiumBy.ACCESSIBILITY_ID, "scanButton")
        scan_btn.click()

        alert = self.driver.switch_to.alert
        assert "camera" in alert.text.lower()
        alert.accept()

    def test_ios_touch_id_login(self):
        """TC-IOS-003: Biometric authentication works on iOS."""
        biometric_btn = self.driver.find_element(AppiumBy.ACCESSIBILITY_ID, "biometricLogin")
        biometric_btn.click()
        time.sleep(2)
        self.driver.execute_script("mobile: enrollBiometric", {"isEnabled": True})
        self.driver.execute_script("mobile: sendBiometricMatch", {"type": "touchId", "match": True})

        home = self.driver.find_element(AppiumBy.ACCESSIBILITY_ID, "homeScreen")
        assert home.is_displayed()

    def test_ios_different_screen_sizes(self):
        """TC-IOS-004: UI adapts to different iPhone screen sizes."""
        window_size = self.driver.get_window_size()
        assert window_size["width"] > 0
        assert window_size["height"] > 0


# ═══════════════════════════════════════════════════════════════════════
# SECTION C: CROSS-DEVICE COMPATIBILITY MATRIX
# ═══════════════════════════════════════════════════════════════════════

class TestCompatibilityMatrix:
    """
    Documents the device compatibility testing matrix.
    These are parameterized test stubs showing the range of devices tested.
    """

    ANDROID_DEVICES = [
        ("Samsung Galaxy S24", "Android 14", "1080x2340"),
        ("Google Pixel 8", "Android 14", "1080x2400"),
        ("OnePlus 12", "Android 14", "1440x3168"),
        ("Samsung Galaxy A54", "Android 13", "1080x2340"),
        ("Xiaomi Redmi Note 13", "Android 13", "1080x2400"),
    ]

    IOS_DEVICES = [
        ("iPhone 15 Pro", "iOS 17", "1179x2556"),
        ("iPhone 14", "iOS 16", "1170x2532"),
        ("iPhone SE 3", "iOS 16", "750x1334"),
        ("iPad Pro 12.9", "iPadOS 17", "2048x2732"),
    ]

    @pytest.mark.parametrize("device,os_version,resolution", ANDROID_DEVICES)
    def test_android_device_compatibility(self, device, os_version, resolution):
        """TC-COMPAT: Verify app works across Android devices."""
        assert device is not None
        assert "Android" in os_version
        width, height = resolution.split("x")
        assert int(width) > 0 and int(height) > 0

    @pytest.mark.parametrize("device,os_version,resolution", IOS_DEVICES)
    def test_ios_device_compatibility(self, device, os_version, resolution):
        """TC-COMPAT: Verify app works across iOS devices."""
        assert device is not None
        assert "iOS" in os_version or "iPad" in os_version
