requirements.txt
# AI Document Enhancement System - Dependencies
# Core Application
PySide6>=6.6.0
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0
pdf2image>=1.16.3
language-tool-python>=2.7.1
textstat>=0.7.3
sumy>=0.11.0
transformers>=4.35.0
torch>=2.1.0
scikit-learn>=1.3.0
python-docx>=1.1.0
fpdf2>=2.7.6

# Testing Tools
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-html>=4.1.0
pytest-qt>=4.2.0
mutmut>=2.4.4
selenium>=4.15.0
Appium-Python-Client>=3.1.0

# Code Quality
flake8>=6.1.0
pylint>=3.0.0
black>=23.10.0
isort>=5.12.0
mypy>=1.6.0
bandit>=1.7.5
safety>=2.3.5

# Build
pyinstaller>=6.1.0
